fix sleep
Waitcnt was in the console display if statement rather than in the main
loop
Time is set with the largest units first (makes more sense, set seconds
closer to the start of the second to minimize error).
Removed references to version numbers and updated the date.
Fixed date not being set on GPS lock