display driver sync to main loop
Lock display buffer 6*refreshRate before the start of the second in
order to sync up the seconds position to the start of the second.
refreshRate should now only be powers of 2 less than 32k in order to
have an integer number of refreshes a second.  There will still be a
short pause (imperceptably small flicker) just before a new second due
to the processing time of the display driver loop.